Move api to pages
