+++ pages/tests/models.py
@@ -2,7 +2,7 @@ from django.contrib.auth.models import AbstractBaseUser, PermissionsMixin
 from feincms.content.richtext.models import RichTextContent
 from user_management.models.mixins import ActiveUserMixin
 
-from pages.api.models import Page
+from pages.models import Page
 
 
 class User(ActiveUserMixin, PermissionsMixin, AbstractBaseUser):
