+++ pages/tests/urls.py
@@ -4,8 +4,8 @@ from django.conf.urls import include, patterns, url
 urlpatterns = patterns('',
     url(r'', include(
         patterns('',
-            url(r'', include('pages.api.urls.page')),
-            url(r'', include('pages.api.urls.group')),
+            url(r'', include('pages.urls.page')),
+            url(r'', include('pages.urls.group')),
         ),
         namespace='pages',
     ))
